\name{pkgDepTools-package}
\alias{pkgDepTools-package}
\alias{pkgDepTools}
\docType{package}
\title{
Package Dependency Tools
}
\description{

  This package provides tools for computing and analyzing dependency
  relationships among R packages.  It provides tools for building a
  graph-based representation of the dependencies among all packages in
  a list of CRAN-style package repositories.  There are also utilities
  for computing installation order of a given package.  If the RCurl
  package is available, an estimate of the download size required to
  install a given package and its dependencies can be obtained.

}
\details{
\tabular{ll}{
Package: \tab pkgDepTools\cr
Type: \tab Package\cr
License: \tab GPL2\cr
}

}
\author{
Seth Falcon
Maintainer: Seth Falcon <sfalcon@fhcrc.org>
}
\keyword{ package }

